Move sidebar link duplicate logic to SidebarLink component
